Frontend: Improve instrument creation and update (#5508)
* rename NewInstrument component to InstrumentForm and update its functionality to support both creating and modifying instruments. Simplify the code. Display more infos about the instrument being updated (name, api, obs_api). Update useless array fields to a string fields to avoid closing select drop down when choosing an option.

* refactor InstrumentForm to use undefined for default values instead of empty strings.

* Update the global style to target empty option from mui select Form and display them grayed out for a better user experience.

* Refactor SourcePublishRelease to simplify release property schema and use the new global style.
